Added type defs and moved functional logic to a seprate file
