+++ package.json
@@ -8,7 +8,6 @@
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.0",
     "typescript": "^4.5.5"
-    
   },
   "scripts": {
     "start": "react-scripts start",
