+++ src/components/signin-with-google-v1/index.d.ts
@@ -1,0 +1,12 @@
+type SignInWithRedirectT = {
+    auth: any,
+    updateUserDetails: Function
+};
+
+type SignInWithPopupT = {
+    auth: any,
+    provider: any,
+    updateUserDetails: Function
+}
+
+export type { SignInWithRedirectT, SignInWithPopupT };
