+++ src/components/signin-with-google-v1/index.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
-import { getAuth, signInWithPopup, GoogleAuthProvider } from "firebase/auth";
+import { getAuth, GoogleAuthProvider } from "firebase/auth";
 import initFirebase from "../../firebase";
+import AuthMethods from "./auth-methods";
 
 const initAuth = (updateUserDetails: Function) => {
 
@@ -11,31 +12,7 @@ const initAuth = (updateUserDetails: Function) => {
     const auth = getAuth();
     const provider = new GoogleAuthProvider();
     // provider.addScope('https://www.googleapis.com/auth/contacts.readonly');
-    signInWithPopup(auth, provider)
-        .then((result) => {
-            // This gives you a Google Access Token. You can use it to access the Google API.
-            // const credential = GoogleAuthProvider.credentialFromResult(result);
-            // const token = credential?.accessToken;
-            // The signed-in user info.
-            const { user } = result;
-            updateUserDetails(user);
-
-        }).catch((error) => {
-            // Handle Errors here.
-            const errorCode = error.code;
-            const errorMessage = error.message;
-            // The email of the user's account used.
-            const email = error.email;
-            // The AuthCredential type that was used.
-            const credential = GoogleAuthProvider.credentialFromError(error);
-            return {
-                errorCode,
-                errorMessage,
-                email,
-                credential,
-            };
-            // ...
-        });
+    AuthMethods.signInWithPopUp({ auth, provider, updateUserDetails });
 }
 
 const SignInWithGoogle = () => {
